requirements.txt
python-telegram-bot==21.6
pytz==2024.2
google-generativeai
